org: reorganizing fluxion as in-tree plugin
This is the start of an experiment to deploy the Flux Framework
scheduler, Fluxion, as an in-tree scheduler plugin. To do this
I am removing the scheduler-plugins repository, but adding it
as a custom in-tree plugin. I am maintaining some of the manifests
from the scheduler-plugins setup that provide the correct rbacs and
services that are needed. This currently builds, and I am at the
step where (finally) I can start designing the internal logic. My
plan is to gut out the scheduler_one.go design, replacing a scheduling
cycle with a simple process that collects groups (where a single pod
can be a group of one). When a group is complete (reached min members
or similar) it will be submit to the fluxion scheduler. If match
allocate returns a result, we pass it to the binding cycle to run
on the cluster. If not, we send to a group queue, which I am going
to implement as a service for now. The WIP groups being assembled
will be stored in a simple map. This is currently for experimental
and testing purposes, and I am hoping to learn a lot and have fun!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>