update: working scheduler
This change has Fluxnetes completely working, on par with Fluence.
The difference is that the design is in-tree, and instead of a custom
CR for PodGroup and a webhook/controller, I remove those two, and use
the default CR for PodGroup from Cosheduling. The ideal design will
not need the group abstraction or can use one from upstream Kubernetes
(a recent KEP was opened for this). I likely will next first work
on some tweaks to the resource spec that I would have wanted for
fluence, and then to harden the generation of the graph. Then my
plan is to slowly add the components I need for the new design
(primarily the map that stores groups being assembled, and new
functions for schedule_one) and then when those are done, I will
swap out the current to test our new ideas. I do not know if it will
work but it will be fun I think.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>