refactor: remove fluence
We do not need most of the fluence code-base with this new design.
I am keeping fluxnetes as an in-tree plugin because it seems like
an optimization to still be able to sort the initial queue of
pods coming in. I am also thinking it would be unwise to completely
remove in-tree plugins - they serve important purposes! But we do
need to more integelligently integrate them with this new design.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>