reservations: add easy with reservation depth 1
This changeset adds reservations to easy, meaning that we
set a reservation depth of 1, and then submit only that
many jobs to ask Flux with the reserve parameter set to
true (or not). Then when a reservation is made, we add to
a new table. The table is then queried later to create
cancel jobs, jobs that are of a new worker type (and queue
within the same strategy) that can be submit to run and
then the reservation table cleared. Now that we have this
additional worker and queue, we also are careful to not
respond to events from it in the main schedule loop. There
is still lots of detail work to be done (and after the core
is here I have a list of tests) but this is really exciting
that it is working so far!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>