cleanup: basic functionality added
This changeset adds support for a duration that drives cleanup,
meaning a duration in seconds can be provided as a label, and then
the label will be populated into the duration (seconds) to kickoff
a cleanup job after allocation. This currently is not doing a cleanup,
as we will need to walk up to a parent level abstraction (often deleting
the pod is not sufficient) and issue cancel to fluxion, but that will
come soon/next. I am also converting the fluxion service container build
to be multi-stage to hopefully make it smaller

Signed-off-by: vsoch <vsoch@users.noreply.github.com>