add owner cleanup
In the case that a duration is set (the pod or pod
template activeDeadlineSeconds) kick off a cleanup
job to remove it from Kubernetes and the scheduler.
Other deletions will need to be handled with events,
which I will do next! I also changed my mind about
the default deletion of 3600 - I think that can be
dangerous for services or other stuffs that should
not be expected to just end. We need to trust the
users creating the abstractions. Also, pods that
are associated with jobs clean themselves up, so
we do not want to double mess with that.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>