queues: queries, cleanup, unschedulable support
This changeset adds a state diagram and the implementation that better moves
single pods through the process of being run through cleanup. I've also added
support for jobs that are not satisfiable - we reject them and then delete
them, which basiscally give them back to the user. We will soon want a
kubectl command that can preserve these final states to present to the
user so they know what happens.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>