cleanup: allow jobs to stay in completed
We currently issue a kubernetes delete / terminate to
all pods up to an owner on termination. This is only
done assuming that we also need to completely delete
the job group for fluxion. Instead, we can check
for if the job pods are all finalized (terminated,
whether successful or failed) before issuing the
fluxion delete. This also allows us to not have
to delete kubernetes objects and allow job objects
to stay in their Completed state for further
inspection

Signed-off-by: vsoch <vsoch@users.noreply.github.com>